requirements.txt
accelerate
altair
autogen-agentchat
autogen-core
autogen-ext
faiss-cpu
filelock
importlib_metadata
ipython
numpy
openai
opencv-python
pandas
pillow
protobuf
psutil
pydantic
pydantic-settings
Pygments
python-dotenv
PyYAML
pyzmq
RapidFuzz
scikit-learn
scipy
sentence-transformers
setuptools
streamlit>=1.33
torch
tqdm
tiktoken>=0.7.0
transformers
typing_extensions
